+++ lib/model/Voter.dart
@@ -14,6 +14,7 @@
  * limitations under the License.
  */
 
+import 'package:rokwire_plugin/utils/datetime_utils.dart';
 import 'package:rokwire_plugin/utils/utils.dart';
 
 class VoterRule {