+++ lib/ui/academics/courses/SkillsScoreChart.dart
@@ -5,6 +5,7 @@ import 'package:flutter/material.dart';
 import 'package:rokwire_plugin/model/survey.dart';
 import 'package:rokwire_plugin/service/localization.dart';
 import 'package:rokwire_plugin/service/styles.dart';
+import 'package:rokwire_plugin/utils/datetime_utils.dart';
 import 'package:rokwire_plugin/utils/utils.dart';
 
 class SkillsScoreChartController {