+++ lib/ui/home/HomeToutWidget.dart
@@ -19,6 +19,7 @@ import 'package:rokwire_plugin/service/notification_service.dart';
 import 'package:rokwire_plugin/service/styles.dart';
 import 'package:rokwire_plugin/ui/panels/modal_image_holder.dart';
 import 'package:rokwire_plugin/ui/widgets/triangle_painter.dart';
+import 'package:rokwire_plugin/utils/datetime_utils.dart';
 import 'package:rokwire_plugin/utils/utils.dart';
 
 class HomeToutWidget extends StatefulWidget {
@@ -281,7 +282,7 @@ class _HomeToutWidgetState extends State<HomeToutWidget> with NotificationsListe
 
   String? get _title1 {
     if (_dayPart != null) {
-      String greeting = AppDateTimeUtils.getDayPartGreeting(dayPart: _dayPart);
+      String greeting = AppRelativeTime.getDayPartGreeting(dayPart: _dayPart);
       if (_firstName?.isNotEmpty ?? false) {
         return "$greeting,";
       }
@@ -323,7 +324,7 @@ class _HomeToutWidgetState extends State<HomeToutWidget> with NotificationsListe
 
   void _updateContent({DayPart? dayPart}) {
     _dayPart = dayPart ?? DateTimeUtils.getDayPart();
-    Storage().homeToutImageUrl = _imageUrl = Content().randomImageUrl('home.tout_1290x750.${dayPartToString(_dayPart)}');
+    Storage().homeToutImageUrl = _imageUrl = Content().randomImageUrl('home.tout_1290x750.${DateTimeUtils.dayPartToString(_dayPart)}');
     Storage().homeToutImageTime = (_imageDateTime = DateTime.now()).millisecondsSinceEpoch;
   }